+++ packages/api-proxy/src/web/api/device/network/onNetworkStatusChange.js
